# src/doc/swagger_doc.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Book List RestAPI",
    "description": "Livros e Listas para todos os gostos!"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "Login de usuário.",
        "parameters": [
          {
            "name": "login",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "example@example.example"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "idUser": {
                      "type": "number",
                      "example": 0
                    },
                    "photo": {
                      "type": "string",
                      "example": "base64 ou null"
                    },
                    "nickname": {
                      "type": "string",
                      "example": "apelido ou null"
                    },
                    "nationality": {
                      "type": "string",
                      "example": "nacionalidade"
                    },
                    "name": {
                      "type": "string",
                      "example": "nome"
                    },
                    "dateOfBirth": {
                      "type": "string",
                      "example": "yyyy-MM-dd"
                    },
                    "email": {
                      "type": "string",
                      "example": "E-mail"
                    },
                    "phone": {
                      "type": "number",
                      "example": 31910101010
                    },
                    "password": {
                      "type": "string",
                      "example": "null"
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/create": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "Cadastro de usuário.",
        "parameters": [
          {
            "name": "create",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "photo": {
                  "type": "string",
                  "example": "base64 ou null"
                },
                "nickname": {
                  "type": "string",
                  "example": "apelido ou null"
                },
                "nationality": {
                  "type": "string",
                  "example": "nacionalidade"
                },
                "name": {
                  "type": "string",
                  "example": "nome"
                },
                "dateOfBirth": {
                  "type": "string",
                  "example": "yyyy-MM-dd"
                },
                "email": {
                  "type": "string",
                  "example": "email@email.email"
                },
                "phone": {
                  "type": "number",
                  "example": 11111111111
                },
                "password": {
                  "type": "string",
                  "example": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucessso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "idUser": {
                      "type": "number",
                      "example": 0
                    },
                    "photo": {
                      "type": "string",
                      "example": "base64 ou null"
                    },
                    "nickname": {
                      "type": "string",
                      "example": "apelido ou null"
                    },
                    "nationality": {
                      "type": "string",
                      "example": "nacionalidade"
                    },
                    "name": {
                      "type": "string",
                      "example": "nome"
                    },
                    "dateOfBirth": {
                      "type": "string",
                      "example": "yyyy-MM-dd"
                    },
                    "email": {
                      "type": "string",
                      "example": "E-mail"
                    },
                    "phone": {
                      "type": "number",
                      "example": 31910101010
                    },
                    "password": {
                      "type": "string",
                      "example": "null"
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/getById": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Buscar usuário pelo ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "Id do usuário a ser buscado.",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "idUser": {
                      "type": "number",
                      "example": 0
                    },
                    "photo": {
                      "type": "string",
                      "example": "base64 ou null"
                    },
                    "nickname": {
                      "type": "string",
                      "example": "apelido ou null"
                    },
                    "nationality": {
                      "type": "string",
                      "example": "nacionalidade"
                    },
                    "name": {
                      "type": "string",
                      "example": "nome"
                    },
                    "dateOfBirth": {
                      "type": "string",
                      "example": "yyyy-MM-dd"
                    },
                    "email": {
                      "type": "string",
                      "example": "E-mail"
                    },
                    "phone": {
                      "type": "number",
                      "example": 31910101010
                    },
                    "password": {
                      "type": "string",
                      "example": "null"
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/update": {
      "put": {
        "tags": [
          "user"
        ],
        "description": "Login de usuário.",
        "parameters": [
          {
            "name": "update",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idUser": {
                  "type": "number",
                  "example": 0
                },
                "photo": {
                  "type": "string",
                  "example": "base64 ou null"
                },
                "nickname": {
                  "type": "string",
                  "example": "apelido ou null"
                },
                "nationality": {
                  "type": "string",
                  "example": "nacionalidade"
                },
                "name": {
                  "type": "string",
                  "example": "nome"
                },
                "dateOfBirth": {
                  "type": "string",
                  "example": "yyyy-MM-dd"
                },
                "email": {
                  "type": "string",
                  "example": "email@email.email"
                },
                "phone": {
                  "type": "number",
                  "example": 11111111111
                },
                "password": {
                  "type": "string",
                  "example": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "idUser": {
                      "type": "number",
                      "example": 0
                    },
                    "photo": {
                      "type": "string",
                      "example": "base64 ou null"
                    },
                    "nickname": {
                      "type": "string",
                      "example": "apelido ou null"
                    },
                    "nationality": {
                      "type": "string",
                      "example": "nacionalidade"
                    },
                    "name": {
                      "type": "string",
                      "example": "nome"
                    },
                    "dateOfBirth": {
                      "type": "string",
                      "example": "yyyy-MM-dd"
                    },
                    "email": {
                      "type": "string",
                      "example": "E-mail"
                    },
                    "phone": {
                      "type": "number",
                      "example": 31910101010
                    },
                    "password": {
                      "type": "string",
                      "example": "null"
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/delete": {
      "delete": {
        "tags": [
          "user"
        ],
        "description": "Login de usuário.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "Id do usuário a ser deletado.",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "Usuário deletado com sucesso."
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  }
}